Refactoring removeSegment flow in upsert
