+++ pinot-integration-tests/src/test/java/org/apache/pinot/integration/tests/models/DummyTableUpsertMetadataManager.java
@@ -107,6 +107,10 @@ public class DummyTableUpsertMetadataManager extends BaseTableUpsertMetadataMana
     }
 
     @Override
+    protected void doRemoveSegment(IndexSegment segment, Iterator<RecordInfo> recordInfoIterator) {
+    }
+
+    @Override
     protected GenericRow doUpdateRecord(GenericRow record, RecordInfo recordInfo) {
       return null;
     }
